moved instructor to new file
